requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-multipart
psycopg2-binary
python-dotenv
numpy
pdfplumber
pdf2image
pytesseract
python-docx
pillow
sentence-transformers
transformers
torch
prometheus-client
pytest
httpx
alembic
cloudinary
